added numeric data type to all currency related columns

+++ db/migration/000001_init_schema.up.sql
@@ -1,7 +1,7 @@
 CREATE TABLE "tasks" (
   "id" BIGSERIAL PRIMARY KEY,
   "taskname" VARCHAR NOT NULL,
-  "budget" FLOAT NOT NULL,
+  "budget" NUMERIC NOT NULL,
   "created_on" timestamptz NOT NULL DEFAULT (now()),
   "created_by" VARCHAR NOT NULL,
   "updated_on" timestamptz NOT NULL DEFAULT (now()),
